epg_flask/requirements.txt
Flask
mysql-connector-python
bcrypt

